chore(tesseract): fold MultiStageAppliedState into QueryProperties
The multi-stage planner used to carry node state in its own type with a
near-duplicate shape of QueryProperties. Now multi-stage state is just a
QueryProperties, and the per-node mutators (set/add dimensions,
filter manipulation, date-range replacements, time-shift accumulation)
live as inherent methods on QueryProperties. Each mutator invalidates
the lazy multi_fact_join_groups cache.

The Rc<MultiStageAppliedState> threading through multi_stage_query_planner
becomes Rc<QueryProperties>; clone_state() becomes a regular owned clone.
Initial state construction goes through the typed builder, so the
no-measures/empty-order semantics are explicit.

Multi-stage CTE deduplication now goes through a dedicated
QueryProperties::eq_as_state that compares the same 7 fields the old
MSAS::PartialEq did — dimensions/time_dimensions (chain-resolved), the
three filter slots, segments and time_shifts — so dedup behaviour is
preserved exactly. The full QueryProperties::PartialEq remains for
query-level equality.

Two pre-existing latent issues fixed in passing:
- change_date_range_filter_impl recursed into nested groups using the
  outer slice instead of group.items — would have looped or silently
  dropped filters on nested groups had any caller produced them.
- input_schema iterated dimensions/time_dimensions through a redundant
  double-clone of Vec; switched to iter().cloned() (after exposing
  state() by reference instead of cloning the Rc).

The orphan PrettyPrint impl on MultiStageAppliedState was the only
content of logical_plan/multistage/common.rs; the file is removed.
